fix: PT-1005 Fix test

+++ order/listener/src/test/kotlin/com/rarible/protocol/order/listener/service/looksrare/LooksrareOrderConverterTest.kt
@@ -69,7 +69,7 @@ internal class LooksrareOrderConverterTest {
         with(orderVersion.data as OrderLooksrareDataV1) {
             assertThat(minPercentageToAsk).isEqualTo(looksrareOrder.minPercentageToAsk)
             assertThat(params).isEqualTo(looksrareOrder.params)
-            assertThat(counter).isEqualTo(looksrareOrder.nonce)
+            assertThat(counter).isEqualTo(looksrareOrder.nonce.toLong())
             assertThat(strategy).isEqualTo(looksrareOrder.strategy)
         }
     }
